feat: "Se corrige funcionamiento de CRUD" // fix: "solucion paginado no funciona despues de uso de modal"
