+++ group-villafañe-mendez/tp-final/src/app/(private)/layout.tsx
@@ -3,7 +3,7 @@
 import type { ReactNode } from "react";
 import { useRouteGuard } from "@/hooks/useRouteGuard";
 import FullScreenLoader from "@/components/loaders/FullScreenLoader";
-import { ProductProvider } from "./dashboard/components/ProductContext";
+import { ProductProvider } from "../../context/ProductContext";
 
 export default function PrivateLayout({ children }: { children: ReactNode }) {
   const { isAuthenticated, isAuthLoading } = useRouteGuard("private");
