+++ group-villafañe-mendez/tp-final/src/components/toasts/Toasts.tsx
@@ -1,33 +1,52 @@
 "use client";
-import React from "react";
+import React, { useEffect, useState } from "react";
+import ReactDOM from "react-dom";
 import { MdErrorOutline, MdCheckCircleOutline } from "react-icons/md";
+import type { ToastType } from "@/context/ToastContext";
 
 interface ToastProps {
   message: string | null;
-  visible: boolean;
-  type?: "error" | "success";
+  type?: ToastType;
+  duration?: number;
 }
 
-export const Toasts = ({ message, visible, type = "error" }: ToastProps) => {
-  if (!message) return null;
+export const Toasts = ({ message, type = "error", duration = 5000 }: ToastProps) => {
+  const [visible, setVisible] = useState(false);
+  const [showDOM, setShowDOM] = useState(false);
+
+  useEffect(() => {
+    if (message) {
+      setShowDOM(true);
+      setVisible(true); 
+      const hideTimer = setTimeout(() => setVisible(false), duration); 
+      const removeTimer = setTimeout(() => setShowDOM(false), duration + 500); 
+      return () => {
+        clearTimeout(hideTimer);
+        clearTimeout(removeTimer);
+      };
+    }
+  }, [message, duration]);
+
+  if (!showDOM) return null;
 
   const bgColor = type === "success" ? "bg-green-600" : "bg-red-900";
   const Icon = type === "success" ? MdCheckCircleOutline : MdErrorOutline;
   const title = type === "success" ? "Success:" : "Error:";
 
-  return (
+  return ReactDOM.createPortal(
     <div
-      className={`fixed bottom-4 right-4 sm:bottom-6 sm:right-6 z-50 flex flex-col items-center p-3 sm:p-4 max-w-xs sm:max-w-sm
+      className={`fixed bottom-4 right-4 sm:bottom-6 sm:right-6 z-[9999] flex flex-col items-center p-3 sm:p-4 max-w-xs sm:max-w-sm
       text-sm sm:text-base text-white 
       ${bgColor} rounded-lg shadow-lg 
       space-y-2
-      transition-all duration-500 ease-in-out
+      transition-opacity transition-transform duration-500 ease-in-out
       ${visible ? "opacity-100 translate-y-0" : "opacity-0 translate-y-4"}`}
     >
       <Icon size={24} />
       <div className="break-words text-center">
         <span className="font-medium">{title}</span> {message}
       </div>
-    </div>
+    </div>,
+    document.body
   );
 };
